+++ src/Digdir.Domain.Dialogporten.WebApi/Common/Authorization/Policy.cs
@@ -1,0 +1,7 @@
+﻿namespace Digdir.Domain.Dialogporten.WebApi.Common.Authorization;
+
+internal static class Policy
+{
+    internal const string Serviceprovider = "serviceprovider";
+    internal const string Enduser = "enduser";
+}
