+++ src/Digdir.Domain.Dialogporten.WebApi/Endpoints/V1/EndUser/Dialog/ListDialogEndpoint.cs
@@ -1,4 +1,5 @@
 ﻿using Digdir.Domain.Dialogporten.Application.Features.V1.EndUser.Dialogs.Queries.List;
+using Digdir.Domain.Dialogporten.WebApi.Common.Extensions;
 using FastEndpoints;
 using MediatR;
