+++ src/Digdir.Domain.Dialogporten.WebApi/Endpoints/V1/EndUser/DialogActivity/GetDialogActivityEndpoint.cs
@@ -1,4 +1,5 @@
 using Digdir.Domain.Dialogporten.Application.Features.V1.EndUser.DialogActivities.Queries.Get;
+using Digdir.Domain.Dialogporten.WebApi.Common.Extensions;
 using FastEndpoints;
 using MediatR;
