+++ src/Digdir.Domain.Dialogporten.WebApi/Endpoints/V1/EndUser/DialogActivity/ListDialogActivityEndpoint.cs
@@ -1,4 +1,5 @@
 using Digdir.Domain.Dialogporten.Application.Features.V1.EndUser.DialogActivities.Queries.List;
+using Digdir.Domain.Dialogporten.WebApi.Common.Extensions;
 using FastEndpoints;
 using MediatR;
