+++ src/Digdir.Domain.Dialogporten.WebApi/Endpoints/V1/EndUser/DialogElement/ListDialogElementEndpoint.cs
@@ -1,4 +1,5 @@
 using Digdir.Domain.Dialogporten.Application.Features.V1.EndUser.DialogElements.Queries.List;
+using Digdir.Domain.Dialogporten.WebApi.Common.Extensions;
 using FastEndpoints;
 using MediatR;
