+++ src/Digdir.Domain.Dialogporten.WebApi/Endpoints/V1/ServiceOwner/DialogElement/CreateDialogElementEndpoint.cs
@@ -5,6 +5,7 @@ using FastEndpoints;
 using MediatR;
 using Medo;
 using IMapper = AutoMapper.IMapper;
+using Digdir.Domain.Dialogporten.WebApi.Common.Extensions;
 
 namespace Digdir.Domain.Dialogporten.WebApi.Endpoints.V1.ServiceOwner.DialogElement;
