+++ src/Digdir.Domain.Dialogporten.WebApi/Endpoints/V1/ServiceOwner/DialogElement/ListDialogElementEndpoint.cs
@@ -1,4 +1,5 @@
 using Digdir.Domain.Dialogporten.Application.Features.V1.ServiceOwner.DialogElements.Queries.List;
+using Digdir.Domain.Dialogporten.WebApi.Common.Extensions;
 using FastEndpoints;
 using MediatR;
