+++ src/Digdir.Domain.Dialogporten.WebApi/Endpoints/V1/ServiceOwner/DialogElement/UpdateDialogElementEndpoint.cs
@@ -3,6 +3,7 @@ using Digdir.Domain.Dialogporten.Application.Features.V1.Common.Localizations;
 using Digdir.Domain.Dialogporten.Application.Features.V1.ServiceOwner.Dialogs.Commands.Update;
 using Digdir.Domain.Dialogporten.Application.Features.V1.ServiceOwner.Dialogs.Queries.Get;
 using Digdir.Domain.Dialogporten.WebApi.Common;
+using Digdir.Domain.Dialogporten.WebApi.Common.Extensions;
 using FastEndpoints;
 using MediatR;
 using IMapper = AutoMapper.IMapper;
