--- src/Digdir.Domain.Dialogporten.WebApi/OptionsSetup/Policy.cs
@@ -1,6 +1,0 @@
-﻿namespace Digdir.Domain.Dialogporten.WebApi.OptionsSetup;
-
-internal static class Policy
-{
-    internal const string Serviceprovider = "serviceprovider";
-}
