+++ src/Digdir.Domain.Dialogporten.WebApi/Program.cs
@@ -1,7 +1,6 @@
 using Digdir.Domain.Dialogporten.Application;
 using Digdir.Domain.Dialogporten.Infrastructure;
 using Digdir.Domain.Dialogporten.Infrastructure.DomainEvents.Outbox.Dispatcher;
-using Digdir.Domain.Dialogporten.WebApi;
 using FastEndpoints;
 using FastEndpoints.Swagger;
 using Microsoft.ApplicationInsights.Extensibility;
@@ -9,9 +8,10 @@ using Serilog;
 using System.Text.Json.Serialization;
 using System.Text.Json.Serialization.Metadata;
 using System.Collections;
-using Digdir.Domain.Dialogporten.WebApi.OptionsSetup;
 using Digdir.Domain.Dialogporten.WebApi.Common.Extensions;
 using Digdir.Domain.Dialogporten.WebApi.Common.Json;
+using Digdir.Domain.Dialogporten.WebApi.Common.OptionsSetup;
+using Digdir.Domain.Dialogporten.WebApi.Common.Options;
 
 // Using two-stage initialization to catch startup errors.
 Log.Logger = new LoggerConfiguration()
